# [SPARK-41171][SQL] Push down filter through window when partitionSpec is empty