--- a/sql/catalyst/src/main/scala/org/apache/spark/sql/catalyst/optimizer/LimitPushDownThroughWindow.scala
+++ b/sql/catalyst/src/main/scala/org/apache/spark/sql/catalyst/optimizer/LimitPushDownThroughWindow.scala
@@ -17,19 +17,21 @@
 
 package org.apache.spark.sql.catalyst.optimizer
 
-import org.apache.spark.sql.catalyst.expressions.{Alias, CurrentRow, DenseRank, IntegerLiteral, NamedExpression, Rank, RowFrame, RowNumber, SpecifiedWindowFrame, UnboundedPreceding, WindowExpression, WindowSpecDefinition}
-import org.apache.spark.sql.catalyst.plans.logical.{Limit, LocalLimit, LogicalPlan, Project, Sort, Window}
+import org.apache.spark.sql.catalyst.expressions.{Alias, CurrentRow, DenseRank, Expression, IntegerLiteral, Literal, NamedExpression, PredicateHelper, Rank, RowFrame, RowNumber, SpecifiedWindowFrame, UnboundedPreceding, WindowExpression, WindowSpecDefinition}
+import org.apache.spark.sql.catalyst.plans.logical.{Filter, Limit, LocalLimit, LogicalPlan, Project, Sort, Window}
 import org.apache.spark.sql.catalyst.rules.Rule
-import org.apache.spark.sql.catalyst.trees.TreePattern.{LIMIT, WINDOW}
+import org.apache.spark.sql.catalyst.trees.TreePattern.{FILTER, LIMIT, WINDOW}
 
 /**
  * Pushes down [[LocalLimit]] beneath WINDOW. This rule optimizes the following case:
  * {{{
  *   SELECT *, ROW_NUMBER() OVER(ORDER BY a) AS rn FROM Tab1 LIMIT 5 ==>
  *   SELECT *, ROW_NUMBER() OVER(ORDER BY a) AS rn FROM (SELECT * FROM Tab1 ORDER BY a LIMIT 5) t
+ *   SELECT *, ROW_NUMBER() OVER(ORDER BY a) AS rn FROM Tab1 WHERE rn <= 5 ==>
+ *   SELECT *, ROW_NUMBER() OVER(ORDER BY a) AS rn FROM (SELECT * FROM Tab1 ORDER BY a LIMIT 5) t
  * }}}
  */
-object LimitPushDownThroughWindow extends Rule[LogicalPlan] {
+object LimitPushDownThroughWindow extends Rule[LogicalPlan] with PredicateHelper {
   // The window frame of RankLike and RowNumberLike can only be UNBOUNDED PRECEDING to CURRENT ROW.
   private def supportsPushdownThroughWindow(
       windowExpressions: Seq[NamedExpression]): Boolean = windowExpressions.forall {
@@ -38,8 +40,24 @@ object LimitPushDownThroughWindow extends Rule[LogicalPlan] {
     case _ => false
   }
 
+  // Extract limit from filter condition and select the min limit.
+  private def extractLimit(
+      condition: Expression,
+      windowExpressions: Seq[NamedExpression]): Option[Int] = {
+    val limits = windowExpressions.collect {
+      case alias @ Alias(WindowExpression(_, _), _) =>
+        extractLimits(condition, alias.toAttribute)
+    }.filter(_.isDefined)
+
+    if (limits.nonEmpty) {
+      Some(limits.flatten.min)
+    } else {
+      None
+    }
+  }
+
   def apply(plan: LogicalPlan): LogicalPlan = plan.transformWithPruning(
-    _.containsAllPatterns(WINDOW, LIMIT), ruleId) {
+    _.containsAnyPattern(WINDOW, LIMIT, FILTER), ruleId) {
     // Adding an extra Limit below WINDOW when the partitionSpec of all window functions is empty.
     case LocalLimit(limitExpr @ IntegerLiteral(limit),
         window @ Window(windowExpressions, Nil, orderSpec, child))
@@ -54,5 +72,17 @@ object LimitPushDownThroughWindow extends Rule[LogicalPlan] {
         limit < conf.topKSortFallbackThreshold =>
       // Sort is needed here because we need global sort.
       project.copy(child = window.copy(child = Limit(limitExpr, Sort(orderSpec, true, child))))
+    case filter @ Filter(condition,
+        window @ Window(windowExpressions, Nil, orderSpec, child))
+      if filter.getTagValue(Filter.FILTER_PUSHED_AS_LIMIT).isEmpty &&
+        supportsPushdownThroughWindow(windowExpressions) && orderSpec.nonEmpty =>
+      val limit = extractLimit(condition, windowExpressions)
+      if (limit.isDefined && limit.get < conf.topKSortFallbackThreshold) {
+        val newWindow = window.copy(child = Limit(Literal(limit.get), Sort(orderSpec, true, child)))
+        filter.setTagValue(Filter.FILTER_PUSHED_AS_LIMIT, ())
+        filter.copy(child = newWindow)
+      } else {
+        filter
+      }
   }
 }