--- a/sql/core/src/test/resources/tpcds-plan-stability/approved-plans-v1_4/q44.sf100/simplified.txt
+++ b/sql/core/src/test/resources/tpcds-plan-stability/approved-plans-v1_4/q44.sf100/simplified.txt
@@ -1,19 +1,19 @@
 TakeOrderedAndProject [rnk,best_performing,worst_performing]
-  WholeStageCodegen (10)
+  WholeStageCodegen (8)
     Project [rnk,i_product_name,i_product_name]
       BroadcastHashJoin [item_sk,i_item_sk]
         Project [rnk,item_sk,i_product_name]
           BroadcastHashJoin [item_sk,i_item_sk]
-            Project [item_sk,rnk,item_sk]
-              BroadcastHashJoin [rnk,rnk]
-                Project [item_sk,rnk]
-                  Filter [rnk,item_sk]
-                    InputAdapter
-                      Window [rank_col]
-                        WholeStageCodegen (3)
-                          Sort [rank_col]
-                            InputAdapter
-                              Exchange #1
+            InputAdapter
+              BroadcastExchange #1
+                WholeStageCodegen (6)
+                  Project [item_sk,rnk,item_sk]
+                    BroadcastHashJoin [rnk,rnk]
+                      Project [item_sk,rnk]
+                        Filter [rnk,item_sk]
+                          InputAdapter
+                            Window [rank_col]
+                              TakeOrderedAndProject [rank_col,item_sk]
                                 WholeStageCodegen (2)
                                   Filter [rank_col]
                                     Subquery #1
@@ -38,23 +38,28 @@ TakeOrderedAndProject [rnk,best_performing,worst_performing]
                                                   ColumnarToRow
                                                     InputAdapter
                                                       Scan parquet spark_catalog.default.store_sales [ss_item_sk,ss_store_sk,ss_net_profit,ss_sold_date_sk]
-                InputAdapter
-                  BroadcastExchange #4
-                    WholeStageCodegen (7)
-                      Project [item_sk,rnk]
-                        Filter [rnk,item_sk]
-                          InputAdapter
-                            Window [rank_col]
-                              WholeStageCodegen (6)
-                                Sort [rank_col]
-                                  InputAdapter
-                                    ReusedExchange [item_sk,rank_col] #1
-            InputAdapter
-              BroadcastExchange #5
-                WholeStageCodegen (8)
-                  Filter [i_item_sk]
-                    ColumnarToRow
                       InputAdapter
-                        Scan parquet spark_catalog.default.item [i_item_sk,i_product_name]
+                        BroadcastExchange #4
+                          WholeStageCodegen (5)
+                            Project [item_sk,rnk]
+                              Filter [rnk,item_sk]
+                                InputAdapter
+                                  Window [rank_col]
+                                    TakeOrderedAndProject [rank_col,item_sk]
+                                      WholeStageCodegen (4)
+                                        Filter [rank_col]
+                                          ReusedSubquery [rank_col] #1
+                                          HashAggregate [ss_item_sk,sum,count] [avg(UnscaledValue(ss_net_profit)),item_sk,rank_col,sum,count]
+                                            InputAdapter
+                                              ReusedExchange [ss_item_sk,sum,count] #2
+            Filter [i_item_sk]
+              ColumnarToRow
+                InputAdapter
+                  Scan parquet spark_catalog.default.item [i_item_sk,i_product_name]
         InputAdapter
-          ReusedExchange [i_item_sk,i_product_name] #5
+          BroadcastExchange #5
+            WholeStageCodegen (7)
+              Filter [i_item_sk]
+                ColumnarToRow
+                  InputAdapter
+                    Scan parquet spark_catalog.default.item [i_item_sk,i_product_name]