… sql/core/src/test/resources/tpcds-plan-stability/approved-plans-v1_4/q44.sf100/explain.txt
… sql/core/src/test/resources/tpcds-plan-stability/approved-plans-v1_4/q44/explain.txt