@beliefer 2023-02-23
https://github.com/apache/spark/pull/40142 replaces this one.